+++ scheduler/scan_scheduler.py
@@ -9,6 +9,8 @@ class ScanScheduler:
 
         self.last_scan = None
 
+        self.next_scan_at = None
+
     def should_scan(self):
 
         if self.last_scan is None:
@@ -31,16 +33,22 @@ class ScanScheduler:
 
         self.last_scan = datetime.now()
 
+        self.next_scan_at = (
+
+            self.last_scan
+
+            + timedelta(
+
+                minutes=Settings.SCAN_INTERVAL_MINUTES
+
+            )
+
+        )
+
     def next_scan_time(self):
 
         if self.last_scan is None:
 
             return "Now"
 
-        next_scan = self.last_scan + timedelta(
-
-            minutes=Settings.SCAN_INTERVAL_MINUTES
-
-        )
-
-        return next_scan.strftime("%I:%M %p")
+        return self.next_scan_at.strftime("%I:%M %p")